+++ spec/controllers/tokens_controller_spec.rb
@@ -90,7 +90,7 @@ describe Doorkeeper::TokensController do
       it 'responds with full token introspection' do
         request.headers['Authorization'] = "Bearer #{access_token.token}"
 
-        post :introspect, token: access_token.token
+        post :introspect, params: { token: access_token.token }
 
         should_have_json 'active', true
         expect(json_response).to include('client_id', 'token_type', 'exp', 'iat')
@@ -104,7 +104,7 @@ describe Doorkeeper::TokensController do
       it 'responds with full token introspection' do
         request.headers['Authorization'] = basic_auth_header_for_client(client)
 
-        post :introspect, token: access_token.token
+        post :introspect, params: { token: access_token.token }
 
         should_have_json 'active', true
         expect(json_response).to include('client_id', 'token_type', 'exp', 'iat')
@@ -119,7 +119,7 @@ describe Doorkeeper::TokensController do
       it 'responds with full token introspection' do
         request.headers['Authorization'] = basic_auth_header_for_client(client)
 
-        post :introspect, token: access_token.token
+        post :introspect, params: { token: access_token.token }
 
         should_have_json 'active', true
         expect(json_response).to include('client_id', 'token_type', 'exp', 'iat')
@@ -135,7 +135,7 @@ describe Doorkeeper::TokensController do
       it 'responds with only active state' do
         request.headers['Authorization'] = basic_auth_header_for_client(different_client)
 
-        post :introspect, token: access_token.token
+        post :introspect, params: { token: access_token.token }
 
         expect(response).to be_successful
 
@@ -151,7 +151,7 @@ describe Doorkeeper::TokensController do
       it 'responds with invalid_client error' do
         request.headers['Authorization'] = basic_auth_header_for_client(client)
 
-        post :introspect, token: access_token.token
+        post :introspect, params: { token: access_token.token }
 
         expect(response).not_to be_successful
         response_status_should_be 401
@@ -168,7 +168,7 @@ describe Doorkeeper::TokensController do
       it 'responds with only active state' do
         request.headers['Authorization'] = basic_auth_header_for_client(client)
 
-        post :introspect, token: SecureRandom.hex(16)
+        post :introspect, params: { token: SecureRandom.hex(16) }
 
         should_have_json 'active', false
         expect(json_response).not_to include('client_id', 'token_type', 'exp', 'iat')
@@ -182,7 +182,7 @@ describe Doorkeeper::TokensController do
       it 'responds with only active state' do
         request.headers['Authorization'] = basic_auth_header_for_client(client)
 
-        post :introspect, token: access_token.token
+        post :introspect, params: { token: access_token.token }
 
         should_have_json 'active', false
         expect(json_response).not_to include('client_id', 'token_type', 'exp', 'iat')
@@ -196,7 +196,7 @@ describe Doorkeeper::TokensController do
       it 'responds with only active state' do
         request.headers['Authorization'] = basic_auth_header_for_client(client)
 
-        post :introspect, token: access_token.token
+        post :introspect, params: { token: access_token.token }
 
         should_have_json 'active', false
         expect(json_response).not_to include('client_id', 'token_type', 'exp', 'iat')
@@ -207,7 +207,7 @@ describe Doorkeeper::TokensController do
       let(:access_token) { FactoryBot.create(:access_token) }
 
       it 'responds with invalid_request error' do
-        post :introspect, token: access_token.token
+        post :introspect, params: { token: access_token.token }
 
         expect(response).not_to be_successful
         response_status_should_be 401